Autofocus textarea on edit and set cursor to end
The cursor is set to the end so the person updating the description in
the chatroom may more quickly add additional information as the
situation unfolds. Rather than having to go to the end of the text
themselves.
The textarea is also autoexpanding so they don't have to drag the
textarea to be larger manually while editing.